requirements.txt
Flask
Flask-Dance
Flask-SQLAlchemy
Flask-Login
Flask-WTF
Werkzeug
pillow
python-dotenv
boto3
psycopg2-binary

